requirements.txt
agate==1.9.1
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
apprise==1.9.4
asgi-lifespan==2.1.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorlover==0.3.0
comm==0.2.2
contourpy==1.3.3
coolname==2.2.0
cryptography==46.0.1
cufflinks==0.17.3
cycler==0.12.1
daff==1.4.2
dateparser==1.2.2
dbt-adapters==1.16.7
dbt-common==1.30.0
dbt-core==1.10.11
dbt-extractor==0.6.0
dbt-postgres==1.9.1
dbt-protos==1.0.375
dbt-semantic-interfaces==0.9.0
debugpy==1.8.15
decorator==5.2.1
deepdiff==8.6.1
docker==7.1.0
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.117.1
fastjsonschema==2.21.2
fonttools==4.59.0
fsspec==2025.9.0
graphviz==0.21
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
leather==0.4.0
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.1
narwhals==2.1.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
oauthlib==3.3.1
opentelemetry-api==1.37.0
orderly-set==5.5.0
orjson==3.11.3
packaging==25.0
pandas==2.3.1
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phpserialize==1.3
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
pluggy==1.6.0
prefect==3.4.19
prometheus_client==0.23.1
prompt_toolkit==3.0.51
protobuf==6.32.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==3.13.0
pyparsing==3.2.3
pytest==8.4.2
pytest-dotenv==0.5.2
pytest-env==1.1.5
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
python-socks==2.7.2
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
readchar==4.2.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.48.0
text-unidecode==1.3
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
typer==0.17.5
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.8.19
uvicorn==0.36.0
wcwidth==0.2.13
websockets==15.0.1
whenever==0.8.8
widgetsnbextension==4.0.14
zipp==3.23.0
